HRMS/backend/requirements.txt
fastapi
uvicorn
boto3
pydantic
email-validator
python-dotenv
pytz
